requirements.txt
# TODO: Figure out how it can only be included in requirements-dev.txt
#django-cors-headers==3.7.0
#drf-nested-routers==0.93.3
#geopy==2.1.0
Babel==2.8.0
django-celery-results==2.0.1
django-celery-beat==2.2.0
Django==3.2.7
#django-filter==2.3.0
#drf_yasg==1.21.4
#djangorestframework==3.14.0
#python-decouple==3.3
python-dotenv==0.14.0
requests==2.24.0
#stripe==2.48.0
#django-meta==1.7.0
#django-social-share==2.0.0
#rest-social-auth==4.2.0
#fcm-django==0.3.4
#django-extensions==3.1.0
django-split-settings==0.3.0
django-settings-export==1.2.1
#django-simple-history==2.11.0
#django-admin-rangefilter==0.6.3
#django-modeltranslation==0.17.2
#pyparsing==2.4.7
django-allauth==0.51.0
#django-utils-six==2.0
django-structlog==4.1.1
#django-pgcrypto-fields==2.6.0
#boto3==1.17.0
#django-storages==1.9.1
#qrcode==7.4.2
Pillow==9.4.0
#curlify==2.2.1
#importlib-metadata==4.13.0
celery==5.2.7
#flower==1.2.0
pandas==1.3.5
celery-progress==0.3
django-celery-results==2.0.1
openpyxl==3.1.2
pyxlsb==1.0.10
django-recaptcha==3.0.0
xlwt==1.3.0
#mysqlclient==2.2.0
xlsxwriter==3.1.6